server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.132.0",
    "langchain>=1.2.10",
    "langchain-experimental>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "matplotlib>=3.10.8",
    "pandas>=3.0.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "seaborn>=0.13.2",
    "uvicorn>=0.41.0",
]
